requirements.txt
datasets
transformers
sentence-transformers
faiss-cpu
huggingface-hub